Phoenix 26.70.0-alpha-2: the beta line can drive a CTRE robot now
The 2.x beta line has been port-and-compile-only since it was cut, because no
Phoenix existed for WPILib 2027 alpha-7 — everything here that runs on a CTRE
drivetrain was compiled and never run. CTRE published 26.70.0-alpha-2 on
2026-09-18, and its changelog says it is "compatible with WPILib 2027 alpha 7 on
the Systemcore on the Alpha/Beta 14 image".

So this line, not alpha-6, is the one that can ship 2.0.0: it is the only one
installable from JitPack, and it is no longer the one that cannot drive a robot.

Three API changes came with the bump, none of them mentioned in the changelog:

- `CANBus.systemcore(int)` and `CANBus.motioncore(int)` are gone. Naming a bus
  is WPILib's job now, through the `org.wpilib.hardware.bus.CANPort` enum, and
  Phoenix takes one in its constructor. `CANPort`'s constants are `CAN_S0`–
  `CAN_S4` and `CAN_D0`–`CAN_D19`, which independently confirms the five- and
  twenty-bus counts `CatalystCANBus` already asserted.
- `setOperatorPerspectiveForward` is now `setOperatorForwardDirection`.
- `CANBusStatus.BusOffCount` and `TxFullCount` widened from `int` to `long`.
  `BusStatus` widens with them rather than casting: a narrowing cast compiles
  and then reports a negative bus-off count on a robot left powered through a
  practice day, which reads as a dashboard bug instead of a lost 64th bit. `REC`
  and `TEC` stay `int` — they are CAN's own 8-bit counters and stop at 255.

Nothing on the wire changes. `CatalystLog.log(String, long)` already existed and
`int` already resolved to it, so `/Catalyst/CAN/Health/<bus>/BusOffCount` keeps
the type both shipped dashboards read.

Verified with `build --no-build-cache --rerun-tasks`, because Gradle has reported
UP-TO-DATE for a classpath it had not rebuilt: 918 tests, 0 failures.

+++ src/main/java/frc/lib/catalyst/hardware/CANBusHealth.java
@@ -48,13 +48,21 @@ public final class CANBusHealth {
 
     private CANBusHealth() {}
 
-    /** A single bus's current state. */
+    /**
+     * A single bus's current state.
+     *
+     * <p>The two cumulative counters are {@code long} because Phoenix 26.70 widened them from
+     * {@code int}. Narrowing them back would compile with a cast and then, on a robot left powered
+     * through a long practice day, report a negative bus-off count — the kind of defect that reads
+     * as a dashboard bug rather than as a lost 64th bit. {@code REC} and {@code TEC} stay
+     * {@code int}: they are CAN's own 8-bit error counters and cannot exceed 255.
+     */
     public record BusStatus(
             String bus,
             boolean ok,
             double utilization,
-            int busOffCount,
-            int txFullCount,
+            long busOffCount,
+            long txFullCount,
             int receiveErrorCount,
             int transmitErrorCount) {
